LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-no-content-value-expected.txt


FAIL Meta color-scheme without content attribute has no effect. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "auto"
